% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{str_split_one}
\alias{str_split_one}
\title{Title}
\usage{
str_split_one(string, patern, n = Inf)
}
\arguments{
\item{string}{A character vector with, at most, one element.}
}
\value{
a character vector
}
\description{
Title
}
\examples{
x <- "alfa,bravo,charlie,delta"
str_split_one(x, patern = ",")
str_split_one(x, patern = ",", n = 2)

y <- "192.168.0.1"
str_split_one(y, patern = stringr::fixed("."))
}
